+++ lib/app.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:focusnest/src/constants/strings.dart';
 import 'package:focusnest/src/routing/app_router.dart';
-import 'package:focusnest/src/services/notfication_controller.dart';
+import 'package:focusnest/src/services/notification_controller.dart';
 import 'package:focusnest/src/utils/theme.dart';
 
 class App extends ConsumerStatefulWidget {
