+++ lib/src/features/activity_timer/presentation/timer_start_screen.dart
@@ -11,7 +11,7 @@ import 'package:focusnest/src/constants/routes_name.dart';
 import 'package:focusnest/src/constants/spacers.dart';
 import 'package:focusnest/src/features/activity_timer/data/activity_timer_database.dart';
 import 'package:focusnest/src/features/activity_timer/data/activity_timer_providers.dart';
-import 'package:focusnest/src/services/notfication_controller.dart';
+import 'package:focusnest/src/services/notification_controller.dart';
 import 'package:focusnest/src/utils/alert_dialogs.dart';
 import 'package:focusnest/src/utils/app_logger.dart';
 import 'package:focusnest/src/utils/date_time_helper.dart';
